Created initial map store for current map states
